Add new piece_ready_alert
Add a new alert, piece_ready_alert, that is generated when a piece has
been verified and written to disk. It is useful when a program needs to
bypass libtorrent and read the piece directly from disk. In my case to
avoid a priority inversion in a realtime streaming application.

Also fixes missing piece_finished_alerts when an exceptions is thrown in
torrent::we_have() before the alert is posted.